Add USE_SASS flag and switch to node-sass
This commit adds a USE_SASS flag to the Makefile
so users can choose to use vanilla CSS and skip
the compilation step if they wish.

Additionally, node-sass is added as an npm dependency
and a helper script for it is added to the build
directory. The helper script is invoked in the Makefile
instead of Ruby Sass.

Lastly, the readme is updated to reflect the above.